requirements.txt
python-dotenv
groq
yfinance